# frontend/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "业务监控系统 API",
        "description": "业务监控系统后端接口文档，包含服务器资产管理、工作流编排、告警配置等功能模块。",
        "contact": {
            "name": "Monitor Team",
            "email": "monitor@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "开发环境"
        }
    ],
...
    "paths": {
...
        "/api/batch/stop/{executionId}": {
            "post": {
                "tags": [
                    "批处理管理"
                ],
                "summary": "停止Job执行",
                "operationId": "stopJob",
                "parameters": [
                    {
                        "name": "executionId",
                        "in": "path",
                        "description": "执行ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/batch/extraction/start": {
            "post": {
                "tags": [
                    "批处理管理"
                ],
                "summary": "启动数据抽取Job",
                "operationId": "startDataExtraction",
                "parameters": [
                    {
                        "name": "cacheKey",
                        "in": "query",
                        "description": "缓存Key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dataSourceId",
                        "in": "query",
                        "description": "数据源ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sql",
                        "in": "query",
                        "description": "SQL语句",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/batch/enrichment/start": {
            "post": {
                "tags": [
                    "批处理管理"
                ],
                "summary": "启动数据富化Job",
                "operationId": "startEnrichment",
                "parameters": [
                    {
                        "name": "cacheKey",
                        "in": "query",
                        "description": "缓存Key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "enrichmentCacheKey",
                        "in": "query",
                        "description": "富化缓存Key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "joinField",
                        "in": "query",
                        "description": "关联字段",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "enrichFields",
                        "in": "query",
                        "description": "富化字段",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "targetTable",
                        "in": "query",
                        "description": "目标表",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/batch/status/{executionId}": {
            "get": {
                "tags": [
                    "批处理管理"
                ],
                "summary": "获取Job执行状态",
                "operationId": "getJobStatus",
                "parameters": [
                    {
                        "name": "executionId",
                        "in": "path",
                        "description": "执行ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/batch/history/{jobName}": {
            "get": {
                "tags": [
                    "批处理管理"
                ],
                "summary": "获取指定Job的运行历史",
                "operationId": "getJobHistory",
                "parameters": [
                    {
                        "name": "jobName",
                        "in": "path",
                        "description": "Job名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "条数限制",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMapStringObject"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "object",
                        "description": "响应数据"
                    }
                },
                "description": "统一API响应"
            },
...
            "ApiResponseMapStringObject": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "object",
                            "description": "响应数据"
                        },
                        "description": "响应数据"
                    }
                },
                "description": "统一API响应"
            },
...
        }
    }
}